graduate cleanup + qol
graduate authed redirect

graduate guest mode auto load plan

remove validation, sidebar cleanup, dto cleanup

tests for dtos and requirement utils

using dal catalog

enum in dto fix + chunks fix
