+++ apps/searchneu/lib/dal/audits.ts
@@ -12,71 +12,8 @@ import {
 } from "../controllers/majors";
 import { auth } from "@/lib/auth/auth";
 import { headers } from "next/headers";
-import { Audit, Requirement, Section } from "../graduate/types";
-
-/**
- * Recursively collects all IRequiredCourse entries from a requirement tree,
- * returning them as "SUBJECT CLASSID" strings.
- */
-function collectRequiredCourseKeys(req: Requirement): string[] {
-  if (req.type === "COURSE") {
-    return [`${req.subject} ${req.classId}`];
-  }
-  if (
-    req.type === "AND" ||
-    req.type === "OR" ||
-    req.type === "XOM" ||
-    req.type === "SECTION"
-  ) {
-    const children =
-      req.type === "SECTION"
-        ? (req as Section).requirements
-        : (req as { courses: Requirement[] }).courses;
-    return children.flatMap(collectRequiredCourseKeys);
-  }
-  return [];
-}
-
-/**
- * Builds a whiteboard pre-populated with courses from the schedule that match
- * requirements in each section. Matching is exact: subject + classId.
- */
-function autoPopulateWhiteboard(
-  sections: Section[],
-  schedule: Audit,
-): Record<string, { courses: string[]; status: string }> {
-  // Collect all "SUBJECT CLASSID" keys from the schedule
-  const scheduleCourseKeys = new Set<string>();
-  for (const year of schedule.years ?? []) {
-    for (const term of [year.fall, year.spring, year.summer1, year.summer2]) {
-      for (const c of term.classes) {
-        scheduleCourseKeys.add(`${c.subject} ${c.classId}`);
-      }
-    }
-  }
-
-  const whiteboard: Record<string, { courses: string[]; status: string }> = {};
-  for (const section of sections) {
-    // Collect every required course key mentioned in this section
-    const sectionCourseKeys = new Set<string>();
-    for (const req of section.requirements) {
-      for (const key of collectRequiredCourseKeys(req)) {
-        sectionCourseKeys.add(key);
-      }
-    }
-
-    // Find which of those appear in the schedule
-    const matched = [...sectionCourseKeys].filter((k) =>
-      scheduleCourseKeys.has(k),
-    );
-
-    whiteboard[section.title] = {
-      courses: matched,
-      status: matched.length > 0 ? "in_progress" : "not_started",
-    };
-  }
-  return whiteboard;
-}
+import { Whiteboard } from "../graduate/types";
+import { buildWhiteboardFromSchedule } from "../graduate/requirementUtils";
 
 /**
  * Verifies the current user from their JWT token and returns their user data.
@@ -156,7 +93,7 @@ export async function createAuditPlan(
 ) {
   const { name, schedule, majors, minors, concentration, catalogYear } =
     createAuditPlanInput;
-  const whiteboard: Record<string, { courses: string[]; status: string }> = {};
+  let whiteboard: Whiteboard = {};
 
   if (majors) {
     const majorName = await getByMajorAndYear(majors, catalogYear ?? 0);
@@ -167,11 +104,10 @@ export async function createAuditPlan(
       );
       return null;
     } else {
-      const populated = autoPopulateWhiteboard(
+      whiteboard = buildWhiteboardFromSchedule(
         majorName.requirementSections,
-        schedule as Audit,
+        schedule,
       );
-      Object.assign(whiteboard, populated);
     }
   }
