--- apps/searchneu/lib/graduate/validation-worker/mock-courses.ts
@@ -1,121 +1,0 @@
-import { course } from "./test-utils";
-const cs2810 = course("CS", 2810);
-const cs3950 = course("CS", 3950);
-const cs4805 = course("CS", 4805);
-const cs4810 = course("CS", 4810);
-const cs4830 = course("CS", 4830);
-const cs4820 = course("CS", 4820);
-const cs4950 = course("CS", 4950, 1);
-const ds3000 = course("DS", 3000);
-const cy4770 = course("CS", 4770);
-const cs1200 = course("CS", 1200, 1);
-const cs1210 = course("CS", 1210, 1);
-const cs1800 = course("CS", 1800);
-const cs1802 = course("CS", 1802, 1);
-const cs2500 = course("CS", 2500);
-const cs2501 = course("CS", 2501, 1);
-const cs2510 = course("CS", 2510);
-const cs2511 = course("CS", 2511, 1);
-const cs2800 = course("CS", 2800);
-const cs2801 = course("CS", 2801, 1);
-const cs3000 = course("CS", 3000);
-const cs3500 = course("CS", 3500);
-const cs3650 = course("CS", 3650);
-const cs3700 = course("CS", 3700);
-const cs3800 = course("CS", 3800);
-const cs4400 = course("CS", 4400);
-const cs4500 = course("CS", 4500);
-const cs4501 = course("CS", 4501);
-const thtr1170 = course("THTR", 1170, 1);
-const cs4410 = course("CS", 4410);
-const cs2550 = course("CS", 2550);
-const ds4300 = course("DS", 4300);
-const cs4300 = course("CS", 4300);
-const math1341 = course("MATH", 1341);
-const math1342 = course("MATH", 1342);
-const math2331 = course("MATH", 2331);
-const math3081 = course("MATH", 3081);
-const phil1145 = course("PHIL", 1145);
-const eece2160 = course("EECE", 2160);
-const chem1211 = course("CHEM", 1211, 3);
-const chem1212 = course("CHEM", 1212, 1);
-const chem1213 = course("CHEM", 1213, 1);
-const chem1214 = course("CHEM", 1214, 3);
-const chem1215 = course("CHEM", 1215, 1);
-const chem1216 = course("CHEM", 1216, 1);
-const phys1151 = course("PHYS", 1151, 3);
-const phys1152 = course("PHYS", 1152, 1);
-const phys1153 = course("PHYS", 1153, 1);
-const phys1155 = course("PHYS", 1155, 3);
-const phys1156 = course("PHYS", 1156, 1);
-const phys1157 = course("PHYS", 1157, 1);
-const engw1111 = course("ENGW", 1111);
-const engw3302 = course("ENGW", 3302);
-const cs1990 = course("CS", 1990);
-const hist1130 = course("HIST", 1130);
-const math2321 = course("MATH", 2321);
-const honr1310 = course("HONR", 1310);
-const math3527 = course("MATH", 3527);
-const artg1250 = course("ARTG", 1250);
-const artg2400 = course("ARTG", 2400);
-export {
-  cs2810,
-  cs3950,
-  cs4805,
-  cs4810,
-  cs4830,
-  cs4820,
-  cs4950,
-  ds3000,
-  cy4770,
-  cs1200,
-  cs1210,
-  cs1800,
-  cs1802,
-  cs2500,
-  cs2501,
-  cs2510,
-  cs2511,
-  cs2800,
-  cs2801,
-  cs3000,
-  cs3500,
-  cs3650,
-  cs3700,
-  cs3800,
-  cs4400,
-  cs4500,
-  cs4501,
-  thtr1170,
-  cs4410,
-  cs2550,
-  ds4300,
-  cs4300,
-  math1341,
-  math1342,
-  math2331,
-  math3081,
-  phil1145,
-  eece2160,
-  chem1211,
-  chem1212,
-  chem1213,
-  chem1214,
-  chem1215,
-  chem1216,
-  phys1151,
-  phys1152,
-  phys1153,
-  phys1155,
-  phys1156,
-  phys1157,
-  engw1111,
-  engw3302,
-  cs1990,
-  hist1130,
-  math2321,
-  honr1310,
-  math3527,
-  artg1250,
-  artg2400,
-};
